refactor(runtime): split ClassRef ToPrimitive out of to_string.rs
#9125 pushed to_string.rs to 2003 lines, over the 2000-line cap
(scripts/check_file_size.sh). Extracted the ClassRef ToPrimitive group
into to_string_class_ref.rs and re-exported class_ref_to_primitive from
to_string so existing crate::value::to_string:: call paths still resolve.
